Split out compose.config.errors
